backend/requirements.txt
fastapi
uvicorn
sqlalchemy
mysql-connector-python
face_recognition
dlib
opencv-python
numpy
pillow
python-multipart
geopy
requests
python-jose[cryptography]
passlib[bcrypt]
python-decouple